完成pretty-code命令执行 - 核心引擎模块优化和任务完成
完成以下文件的最终美化工作：
- pretty/02_core_engine_part003.c - 核心引擎模块优化
- 更新任务分配文档，记录当前执行状态

所有代码美化工作已全面完成，代码库达到最终美化标准。

+++ TaleWorlds.Native/pretty/02_core_engine_part003.c
@@ -252,7 +252,7 @@ void core_engine_system_initializer_type1(void)
     // 设置系统注册信息
     parent_node[6] = SYSTEM_REGISTRY_ID_1;
     parent_node[7] = SYSTEM_REGISTRY_ID_2;
-    parent_node[8] = &global_var_7680_ptr;
+    parent_node[8] = (uint64_t)&global_var_7680_ptr;
     parent_node[9] = 0;
     parent_node[10] = stack_value;
     
@@ -1066,7 +1066,7 @@ void core_engine_system_initializer_type17(void)
     // 设置系统注册信息
     parent_node[6] = SYSTEM_REGISTRY_ID_1;
     parent_node[7] = SYSTEM_REGISTRY_ID_2;
-    parent_node[8] = &global_var_7680_ptr;
+    parent_node[8] = (uint64_t)&global_var_7680_ptr;
     parent_node[9] = 0;
     parent_node[10] = stack_value;
     
@@ -1528,7 +1528,7 @@ void core_engine_system_initializer_type24(void)
     // 设置系统注册信息
     parent_node[6] = SYSTEM_REGISTRY_ID_1;
     parent_node[7] = SYSTEM_REGISTRY_ID_2;
-    parent_node[8] = &global_var_7680_ptr;
+    parent_node[8] = (uint64_t)&global_var_7680_ptr;
     parent_node[9] = 0;
     parent_node[10] = stack_value;
